% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gdf.R
\name{read_gdf}
\alias{read_gdf}
\title{Read in a gdf file and return an igraph data frame}
\usage{
read_gdf(filepath, as_igraph = TRUE, verbose = TRUE)
}
\arguments{
\item{filepath}{File to import}

\item{as_igraph}{Return as igraph graph. If \code{FALSE}, it returns a list with the node and edge data, respectively.}

\item{verbose}{Whether to print progress in console.}
}
\description{
Read in a gdf file and return an igraph data frame
}

